hooks/configure: ensure configuration updates
Some logic is borrowed from ubuntu-frame; generally, a backup and a
current config will be saved to $SNAP_COMMON and the server should
automatically restart if the config changes.

If a config is not set (via a gadget snap or is just outright unset), we
avoid failure by making sure the server is disabled.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>